+++ core/src/elements/ons-popover/index.js
@@ -235,14 +235,6 @@ class PopoverElement extends BaseElement {
     this._animator = (options) => factory.newAnimator(options);
   }
 
-  _onDeviceBackButton(event) {
-    if (this.cancelable) {
-      this._cancel();
-    } else {
-      event.callParentHandler();
-    }
-  }
-
   _positionPopover(target) {
     const {_radius: radius, _content: el, _margin: margin} = this;
     const pos = target.getBoundingClientRect();
@@ -526,18 +518,29 @@ class PopoverElement extends BaseElement {
 
   /**
    * @property onDeviceBackButton
-   * @readonly
    * @type {Object}
    * @description
-   *   [en]Retrieve the back-button handler.[/en]
-   *   [ja]バックボタンハンドラを取得します。[/ja]
+   *   [en]Back-button handler.[/en]
+   *   [ja]バックボタンハンドラ。[/ja]
    */
   get onDeviceBackButton() {
     return this._backButtonHandler;
   }
 
+  set onDeviceBackButton(callback) {
+    if (this._backButtonHandler) {
+      this._backButtonHandler.destroy();
+    }
+
+    this._backButtonHandler = deviceBackButtonDispatcher.createHandler(this, callback);
+  }
+
+  _resetBackButtonHandler() { // do we need this twice?
+    this.onDeviceBackButton = e => this.cancelable ? this._cancel() : e.callParentHandler();
+  }
+
   attachedCallback() {
-    this._backButtonHandler = deviceBackButtonDispatcher.createHandler(this, this._onDeviceBackButton.bind(this));
+    this._resetBackButtonHandler();
 
     contentReady(this, () => {
       this._margin = this._margin || parseInt(window.getComputedStyle(this).getPropertyValue('top'));
@@ -545,7 +548,7 @@ class PopoverElement extends BaseElement {
 
       this._mask.addEventListener('click', this._boundCancel, false);
 
-      this._backButtonHandler = deviceBackButtonDispatcher.createHandler(this, this._onDeviceBackButton.bind(this));
+      this._resetBackButtonHandler();
 
       this._popover.addEventListener('DOMNodeInserted', this._boundOnChange, false);
       this._popover.addEventListener('DOMNodeRemoved', this._boundOnChange, false);